% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCGP.R
\name{fastCGP}
\alias{fastCGP}
\title{fastCGP}
\usage{
fastCGP(snp2gene_file, snp_chr_pos_p_file, genes2compute_file)
}
\arguments{
\item{snp2gene_file}{(the path of) a space-delimited two-column text file that contains information of which SNPs are mapped to which genes. The column headers of this file should be exactly gene and SNP (order insensitive)}

\item{snp_chr_pos_p_file}{a space-delimited four-column text file that contains the information of all SNPs that are analyzed in a GWAS, including the name of a SNP, its chromosomal number, position on the chromosome, and GWAS association P-value. The column headers of this file should be exactly SNP, chr, pos, and p (order insensitive). More SNPs contained in this file will lead to better precision of the result}

\item{genes2compute_file}{(optional): (the path of) a one-column text file that contains the list of genes which you want to compute their P-values. The column header of this file should be exactly gene. If this file is not provided, all genes included in the snp_chr_pos_p_file will be computed}
}
\description{
Compute gene level p-value from SNP p-values using fast circular genomic permutation
}
